Create RADME.md
HiRAM, the GFDL global HIgh Resolution Atmospheric Model, was developed with a goal of providing an improved representation of significant weather events in a global climate model. Our intention was to produce a model capable of simulating the statistics of tropical storms, with sufficient fidelity that it can be used with confidence to study the causes of year-to-year variability in storm activity, recent trends in activity, as well as the predictability of the Atlantic hurricane season. As the credibility of the model improves, based on comparisons with observations, we will apply it to study the effects of global warming on tropical storms.

HiRAM was developed based on AM2 (GAMDT 2004) with increased horizontal and vertical resolutions, as well as simplified parameterizations for moist convection and large-scale (stratiform) cloudiness. The idea behind the simplifications in physics is to make the parameterized convection less intrusive so that explicit convection can make significant contributions to the vertical transport of water and energy in the tropics. The idea of the parameterized convection helping the resolved scale convection, rather than supplanting it entirely, is a strategy that we believe will be useful in working with global models with finer and finer resolution, as the resolved convection begins to take on some realistic features (Pauluis and Garner, 2006).

Specific changes from AM2 include: 1) the finite-volume dynamical core (Lin 2004) on a latitude-longitude grid has been replaced by a finite-volume core using a cubed-sphere grid topology (Putman and Lin 2007) with a quasi-uniform horizontal grid spacing; 2) the number of vertical levels has been increased from 24 to 32; 3) the relaxed Arakawa-Schubert convective closure (Moorthi and Suarez 1992) in AM2 has been replaced by a scheme based on the parameterization of shallow convection by Bretherton et al. (2004); and 4) the prognostic cloud fraction scheme has been replaced by a simpler diagnostic scheme assuming a sub-grid scale distribution of total water.

HiRAM retains the surface flux, boundary layer, land surface, gravity wave drag, large-scale cloud microphysics, and radiative transfer modules from AM2. See Zhao et al. (2009) for more details about the modifications.

At about 50km horizontal grid size, the HIRAM (C180-HIRAM) forced by the observed ocean surface temperatures is found to be able to simulate many aspects of the observed tropical cyclone frequency variability for the past few decades, for which reliable observations are available (Zhao et al. 2009). These include the geographical distribution of global hurricane tracks, the seasonal cycle, as well as the inter-annual variability and the decadal trend of hurricane frequency over multiple ocean basins.

HiRAM has been used to study hurricane seasonal forecast in the N. Atlantic (Zhao et al. 2010, Chen and Lin 2011) and the results support the view that the overall activity of the Atlantic hurricane season has substantial predictability, if we can predict ocean temperatures. These studies also motivate the development of the GFDL Hybrid Hurricane Forecast System (HyHUFS, Vecchi et al. 2011).

HiRAM has also been used to study the response of tropical storms to global warming and CO2 increase (Held and Zhao 2011, Zhao and Held 2010, 2012). HIRAM is also a GFDL model currently participating in the CMIP5 high resolution time-slice simulations (Held et al., in preparation) and the US CLIVAR Hurricane Working Group.

The HiRAM atmospheric model is coupled to the new land model LM3, shared by all of the GFDL models contributed to CMIP5.